refactor: move community gradient logic to dedicated utils file
